refactor: create exec.pnpm-cli-runner
