+++ .changeset/fifty-emus-fold.md
@@ -1,0 +1,5 @@
+---
+"@pnpm/exec.pnpm-cli-runner": major
+---
+
+Initial release.
