+++ exec/pnpm-cli-runner/README.md
@@ -1,0 +1,15 @@
+# @pnpm/exec.pnpm-cli-runner
+
+> Runs pnpm CLI
+
+[![npm version](https://img.shields.io/npm/v/@pnpm/exec.pnpm-cli-runner.svg)](https://www.npmjs.com/package/@pnpm/exec.pnpm-cli-runner)
+
+## Installation
+
+```sh
+pnpm add @pnpm/exec.pnpm-cli-runner
+```
+
+## License
+
+MIT
