+++ exec/pnpm-cli-runner/package.json
@@ -1,0 +1,42 @@
+{
+  "name": "@pnpm/exec.pnpm-cli-runner",
+  "version": "1000.0.0-0",
+  "description": "Runs pnpm CLI",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "files": [
+    "lib",
+    "!*.map"
+  ],
+  "engines": {
+    "node": ">=18.12"
+  },
+  "scripts": {
+    "lint": "eslint \"src/**/*.ts\"",
+    "test": "pnpm run compile",
+    "prepublishOnly": "pnpm run compile",
+    "compile": "tsc --build && pnpm run lint --fix"
+  },
+  "repository": "https://github.com/pnpm/pnpm/blob/main/exec/pnpm-cli-runner",
+  "keywords": [
+    "pnpm10"
+  ],
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/pnpm/pnpm/issues"
+  },
+  "homepage": "https://github.com/pnpm/pnpm/blob/main/exec/pnpm-cli-runner#readme",
+  "devDependencies": {
+    "@pnpm/exec.pnpm-cli-runner": "workspace:*"
+  },
+  "dependencies": {
+    "execa": "catalog:"
+  },
+  "funding": "https://opencollective.com/pnpm",
+  "exports": {
+    ".": "./lib/index.js"
+  },
+  "jest": {
+    "preset": "@pnpm/jest-config"
+  }
+}
